agent-driven: refactor(datetime): Move Trio definition back to semantic_skeletons.rs
