+++ provider/source/src/datetime/semantic_skeletons.rs
@@ -2,7 +2,7 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
-use super::{DatagenCalendar, PackedPatternItem, Trio, select_pattern, transpose_with_fallback};
+use super::{DatagenCalendar, PackedPatternItem, select_pattern, transpose_with_fallback};
 use crate::debug_provider::DebugProvider;
 use crate::{IterableDataProviderCached, SourceDataProvider, cldr_serde};
 use icu::datetime::fieldsets::enums::*;
@@ -21,6 +21,39 @@ use std::collections::{BTreeMap, HashSet};
 
 #[cfg(test)]
 mod tests;
+
+/// A generic structure holding a standard pattern and two optional variants.
+/// See `GenericPackedPatterns` for more details on variants.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(crate) struct Trio<T> {
+    pub standard: T,
+    pub variant0: Option<T>,
+    pub variant1: Option<T>,
+}
+
+impl<T> Trio<T> {
+    /// Returns an iterator over mutable references in quality order (based on a key function).
+    pub fn iter_in_quality_order_mut<K: Ord>(
+        &mut self,
+        mut key_fn: impl FnMut(&T) -> K,
+    ) -> impl Iterator<Item = &mut T> {
+        let mut list = [
+            Some(&mut self.standard),
+            self.variant0.as_mut(),
+            self.variant1.as_mut(),
+        ];
+        list.sort_by_key(|variant| variant.as_ref().map(|v| key_fn(*v)));
+        list.into_iter().flatten()
+    }
+
+    pub fn map<U>(self, mut f: impl FnMut(T) -> U) -> Trio<U> {
+        Trio {
+            standard: f(self.standard),
+            variant0: self.variant0.map(&mut f),
+            variant1: self.variant1.map(&mut f),
+        }
+    }
+}
 
 /// Some patterns associated with a [`SkeletonQuality`].
 #[derive(Debug, Clone, PartialEq)]
